+++ apps/registry/src/layouts/Layout.astro
@@ -1,5 +1,5 @@
 ---
-import "~/styles/global.css";
+import "#styles";
 
 interface Props {
   title: string;
@@ -26,12 +26,10 @@ const { title, description } = Astro.props;
     <title>{title}</title>
     <meta name="description" content={description} />
 
-    <!-- Open Graph -->
     <meta property="og:title" content={title} />
     <meta property="og:description" content={description} />
     <meta property="og:url" content={canonical} />
 
-    <!-- Twitter -->
     <meta name="twitter:title" content={title} />
     <meta name="twitter:description" content={description} />
