+++ apps/registry/src/pages/500.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from "~/layouts/Layout.astro";
+import Layout from "#layout";
 
 export const prerender = true;
