+++ apps/registry/src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
-import TopBar from "~/components/TopBar.astro";
-import Layout from "~/layouts/Layout.astro";
+import TopBar from "#components/TopBar.astro";
+import Layout from "#layout";
 
 export const prerender = true;
