+++ apps/registry/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "extends": "astro/tsconfigs/strictest",
   "compilerOptions": {
-    "paths": { "~/*": ["./src/*"] },
     "jsxImportSource": "solid-js",
     "jsx": "preserve"
   },
