+++ apps/registry/wrangler.jsonc
@@ -3,7 +3,7 @@
 
   "name": "u",
 
-  "compatibility_date": "2025-12-02",
+  "compatibility_date": "2026-03-17",
 
   "compatibility_flags": ["nodejs_compat"],
