--- biome.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "https://biomejs.dev/schemas/2.4.8/schema.json",
-  "extends": ["@gameroman/config/biome"]
-}
