+++ biome.jsonc
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://biomejs.dev/schemas/2.4.9/schema.json",
+  "extends": ["@gameroman/config/biome"]
+}
