+++ package.json
@@ -7,15 +7,15 @@
       "packages/*"
     ],
     "catalog": {
-      "@astrojs/cloudflare": "^13.1.3",
+      "@astrojs/cloudflare": "^13.1.4",
       "@astrojs/solid-js": "^6.0.1",
       "@tailwindcss/vite": "^4.2.2",
-      "astro": "^6.0.8",
-      "rolldown": "^1.0.0-rc.9",
-      "solid-js": "^1.9.11",
+      "astro": "^6.1.1",
+      "rolldown": "^1.0.0-rc.12",
+      "solid-js": "^1.9.12",
       "tailwindcss": "^4.2.2",
-      "tsdown": "^0.21.4",
-      "wrangler": "^4.76.0",
+      "tsdown": "^0.21.7",
+      "wrangler": "^4.78.0",
       "zod": "^4.3.6"
     }
   },
@@ -28,8 +28,8 @@
     "test": "bun run --cwd ./packages/usts test"
   },
   "devDependencies": {
-    "@biomejs/biome": "2.4.8",
-    "@gameroman/config": "^0.0.2"
+    "@biomejs/biome": "2.4.9",
+    "@gameroman/config": "^0.0.5"
   },
   "overrides": {
     "esbuild": "^0.27.3",
@@ -39,6 +39,8 @@
     "@babel/types": "8.0.0-rc.3",
     "@babel/traverse": "8.0.0-rc.3",
     "@babel/helper-validator-identifier": "8.0.0-rc.3",
-    "unstorage": "alpha"
+    "unstorage": "alpha",
+    "picomatch": "latest",
+    "@cloudflare/vite-plugin": "latest"
   }
 }
