+++ packages/usts/tsconfig.json
@@ -1,31 +1,8 @@
 {
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "declaration": true,
-    "isolatedDeclarations": true,
-
-    "target": "esnext",
-    "module": "esnext",
-    "moduleResolution": "bundler",
-    "moduleDetection": "force",
-
-    "rootDir": "./",
-
-    "paths": { "~/*": ["./src/*"] },
-
-    "skipLibCheck": true,
-
-    // Max strictness
-    "strict": true,
-    "noFallthroughCasesInSwitch": true,
-    "noUncheckedIndexedAccess": true,
-    "noImplicitOverride": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noPropertyAccessFromIndexSignature": true,
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "forceConsistentCasingInFileNames": true,
-    "verbatimModuleSyntax": true
+    "isolatedDeclarations": true
   },
   "include": ["src", "tests"]
 }
