+++ packages/usts/tsdown.config.ts
@@ -4,11 +4,12 @@ export default defineConfig({
   entry: {
     core: "src/core/index.ts",
     config: "src/config/index.ts",
-    "bin/cli": "src/bin/cli.ts",
+    cli: "src/cli/index.ts",
   },
   copy: { from: "src/types/*.d.ts", to: "dist" },
   exports: {
     enabled: true,
+    bin: true,
     customExports(pkg) {
       pkg["./types"] = "./dist/virtual.d.ts";
       return pkg;
